requirements.txt
beautifulsoup4==4.11.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
cssselect==1.2.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
html5lib==1.1
httplib2==0.21.0
idna==3.4
joblib==1.2.0
lxml==4.9.2
msgpack==1.0.4
numpy==1.24.2
pandas==1.5.3
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
sportsipy==0.6.0
sportsreference==0.5.2
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.14
webencodings==0.5.1
